vittusaatana

+++ bankautomat/apusivu.cpp
@@ -6,9 +6,10 @@ QTimer *olioTyyppiQtimer = new QTimer;
 QTimer *olioPinQtimer = new QTimer;
 QTimer *olioToimintosivuQtimer = new QTimer;
 QTimer *olioNostaRahaaQtimer = new QTimer;
+//olioNostaRahaaQtimer = new QTimer;
 QTimer *olioPanoQtimer = new QTimer;
 
-short timerCounter = 0;
-short timerCounter2 = 0;
+int timerCounter = 0;
+int timerCounter2 = 0;
 short timerAika1 = 5;
 short timerAika2 = 50;
